Move extract_secret onto the transaction
We need to do this so we aren't forced to implement it for all
ledgers. Instead we just implement LQS events for all ledgers where
the transaction is ExtractSecret.
