+++ application/comit_node/src/swap_protocols/rfc003/actions/bob/btc_eth.rs
@@ -98,10 +98,10 @@ impl StateActions for SwapStates<Bob<Bitcoin, Ethereum, BitcoinQuantity, EtherQu
             SS::AlphaFundedBetaRedeemed(AlphaFundedBetaRedeemed {
                 ref swap,
                 ref alpha_htlc_location,
-                ref secret,
+                ref beta_redeemed_tx,
                 ..
             }) => vec![Action::Redeem(
-                swap.redeem_action(*alpha_htlc_location, *secret),
+                swap.redeem_action(*alpha_htlc_location, beta_redeemed_tx.secret),
             )],
             _ => vec![],
         }
