+++ application/comit_node/src/swap_protocols/rfc003/ledger.rs
@@ -23,29 +23,25 @@ pub trait Ledger: swap_protocols::Ledger {
         + PartialEq
         + Debug
         + Into<<Self as swap_protocols::ledger::Ledger>::Identity>;
+}
 
-    fn extract_secret(
-        transaction: &RedeemTransaction<Self>,
-        secret_hash: &SecretHash,
-    ) -> Option<Secret>;
+pub trait ExtractSecret {
+    fn extract_secret(&self, secret_hash: &SecretHash) -> Option<Secret>;
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct FundTransaction<L: Ledger>(pub L::Transaction);
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-pub struct RedeemTransaction<L: Ledger>(pub L::Transaction);
+pub struct RedeemTransaction<L: Ledger> {
+    pub transaction: L::Transaction,
+    pub secret: Secret,
+}
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct RefundTransaction<L: Ledger>(pub L::Transaction);
 
 impl<L: Ledger> AsRef<L::Transaction> for FundTransaction<L> {
-    fn as_ref(&self) -> &L::Transaction {
-        &self.0
-    }
-}
-
-impl<L: Ledger> AsRef<L::Transaction> for RedeemTransaction<L> {
     fn as_ref(&self) -> &L::Transaction {
         &self.0
     }
