+++ application/comit_node/src/swap_protocols/rfc003/mod.rs
@@ -24,7 +24,7 @@ mod state_machine_test;
 
 pub use self::{
     error::Error,
-    ledger::{FundTransaction, Ledger, RedeemTransaction, RefundTransaction},
+    ledger::{ExtractSecret, FundTransaction, Ledger, RedeemTransaction, RefundTransaction},
     save_state::SaveState,
     secret::{RandomnessSource, Secret, SecretFromErr, SecretHash},
 };
